pyproject.toml
[tool.poetry]
name = "patherea-p2p"
version = "0.1.0"
description = "Point-Supervised Cell Detection using P2P."
authors = ["Dejan Stepec <dejan.stepec@patherea.ai>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "~3.9"
loguru = "^0.6.0"
albumentations = "^1.3.0"
imgaug = "^0.4.0"
lightning-flash = "^0.8.1.post0"
numba = "^0.56"
hydra-core = "^1.3.2"
mlflow = "^2.1.1"
hydra-optuna-sweeper = "^1.2.0"
hydra-submitit-launcher = "^1.2.0"
h5py = "^3.9.0"
timm = "0.4.12"
mmcv-full = "1.4.2"
xformers = "^0.0.27"
ninja = "^1.11.1.1"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black = "^24.8.0"
isort = "^5.11.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
